… tests/files/grayscale_2_should_be_grayscale_1.png
… tests/files/grayscale_4_should_be_grayscale_1.png
… tests/files/grayscale_4_should_be_grayscale_2.png
… tests/files/grayscale_8_should_be_grayscale_1.png
… tests/files/grayscale_8_should_be_grayscale_2.png
… tests/files/grayscale_8_should_be_grayscale_4.png
… tests/reduction.rs