This PR started out as some initial changes to prepare for a raw api but I ended up down a rabbit hole when I noticed grayscale reductions were quite slow.

- Changed PngImage.data and ScanLines to work without filter bytes, as mentioned [here](https://github.com/shssoichiro/oxipng/issues/376#issuecomment-1349915324). This allows some simplifications in the reductions as they don't have to work with filter bytes and in most cases can work directly with png.data instead of scan lines.
- Reworked `reduce_rgb_to_grayscale` to be much faster. The separate `reduce_rgba_to_grayscale_alpha` is no longer needed.
- Reworked `reduce_color_type` to be much easier to read. Moved RGBA alpha reduction after the palette reduction so it can be skipped if palette reduction was successful.
- Reworked `reduce_bit_depth_8_or_less` to be much faster by using bitwise logic instead of bitvecs. This now also allows grayscale reductions from 4 or 2 bits (previously only worked for 8 bits).
- Added tests and benches for grayscale depth reductions.

Before:
```
test reductions_16_to_8_bits                ... bench:   1,160,455 ns/iter (+/- 87,440)
test reductions_2_to_1_bits                 ... bench:   1,517,861 ns/iter (+/- 61,723)
test reductions_4_to_1_bits                 ... bench:   2,104,311 ns/iter (+/- 60,148)
test reductions_4_to_2_bits                 ... bench:   3,001,932 ns/iter (+/- 99,429)
test reductions_8_to_1_bits                 ... bench:   3,499,886 ns/iter (+/- 105,868)
test reductions_8_to_2_bits                 ... bench:   4,201,524 ns/iter (+/- 135,940)
test reductions_8_to_4_bits                 ... bench:   5,976,396 ns/iter (+/- 151,568)
test reductions_alpha                       ... bench:     211,101 ns/iter (+/- 24,003)
test reductions_grayscale_2_to_1_bits       ... bench:   2,396,057 ns/iter (+/- 70,238)
test reductions_grayscale_4_to_1_bits       ... bench:   4,818,521 ns/iter (+/- 111,236)
test reductions_grayscale_4_to_2_bits       ... bench:   4,821,135 ns/iter (+/- 90,225)
test reductions_grayscale_8_to_1_bits       ... bench:   3,748,953 ns/iter (+/- 96,778)
test reductions_grayscale_8_to_2_bits       ... bench:   4,892,843 ns/iter (+/- 644,695)
test reductions_grayscale_8_to_4_bits       ... bench:   8,288,725 ns/iter (+/- 238,924)
test reductions_palette_duplicate_reduction ... bench:     489,503 ns/iter (+/- 66,429)
test reductions_palette_full_reduction      ... bench:     490,014 ns/iter (+/- 43,776)
test reductions_palette_unused_reduction    ... bench:     480,204 ns/iter (+/- 22,962)
test reductions_rgb_to_grayscale_16         ... bench:  71,380,260 ns/iter (+/- 5,558,108)
test reductions_rgb_to_grayscale_8          ... bench:  41,314,202 ns/iter (+/- 1,293,874)
test reductions_rgb_to_palette_8            ... bench:   2,412,796 ns/iter (+/- 133,771)
test reductions_rgba_to_grayscale_16        ... bench:  59,053,749 ns/iter (+/- 1,216,070)
test reductions_rgba_to_grayscale_8         ... bench:  42,467,924 ns/iter (+/- 983,057)
test reductions_rgba_to_grayscale_alpha_16  ... bench:  57,852,657 ns/iter (+/- 1,990,585)
test reductions_rgba_to_grayscale_alpha_8   ... bench:  29,321,227 ns/iter (+/- 730,549)
test reductions_rgba_to_palette_8           ... bench:   3,864,159 ns/iter (+/- 95,834)
test reductions_rgba_to_rgb_16              ... bench:   1,696,329 ns/iter (+/- 74,873)
test reductions_rgba_to_rgb_8               ... bench:   1,454,472 ns/iter (+/- 52,481)
```

After
```
test reductions_16_to_8_bits                ... bench:   1,046,212 ns/iter (+/- 39,313)
test reductions_2_to_1_bits                 ... bench:     292,724 ns/iter (+/- 21,898)
test reductions_4_to_1_bits                 ... bench:     314,482 ns/iter (+/- 11,063)
test reductions_4_to_2_bits                 ... bench:     365,989 ns/iter (+/- 525,603)
test reductions_8_to_1_bits                 ... bench:     365,478 ns/iter (+/- 29,957)
test reductions_8_to_2_bits                 ... bench:     390,015 ns/iter (+/- 30,270)
test reductions_8_to_4_bits                 ... bench:     447,851 ns/iter (+/- 44,147)
test reductions_alpha                       ... bench:     219,393 ns/iter (+/- 402,640)
test reductions_grayscale_2_to_1_bits       ... bench:     313,803 ns/iter (+/- 21,367)
test reductions_grayscale_4_to_1_bits       ... bench:     368,708 ns/iter (+/- 4,076)
test reductions_grayscale_4_to_2_bits       ... bench:     429,138 ns/iter (+/- 15,646)
test reductions_grayscale_8_to_1_bits       ... bench:     474,770 ns/iter (+/- 57,106)
test reductions_grayscale_8_to_2_bits       ... bench:     568,581 ns/iter (+/- 54,894)
test reductions_grayscale_8_to_4_bits       ... bench:     799,713 ns/iter (+/- 23,580)
test reductions_palette_duplicate_reduction ... bench:     173,649 ns/iter (+/- 518,764)
test reductions_palette_full_reduction      ... bench:     166,232 ns/iter (+/- 13,515)
test reductions_palette_unused_reduction    ... bench:     171,974 ns/iter (+/- 20,264)
test reductions_rgb_to_grayscale_16         ... bench:   1,067,927 ns/iter (+/- 27,111)
test reductions_rgb_to_grayscale_8          ... bench:   1,066,505 ns/iter (+/- 50,378)
test reductions_rgb_to_palette_8            ... bench:   2,453,904 ns/iter (+/- 104,640)
test reductions_rgba_to_grayscale_16        ... bench:   2,869,405 ns/iter (+/- 75,703)
test reductions_rgba_to_grayscale_8         ... bench:   2,442,731 ns/iter (+/- 71,920)
test reductions_rgba_to_grayscale_alpha_16  ... bench:   1,248,387 ns/iter (+/- 70,679)
test reductions_rgba_to_grayscale_alpha_8   ... bench:   1,078,038 ns/iter (+/- 66,346)
test reductions_rgba_to_palette_8           ... bench:   2,199,989 ns/iter (+/- 64,124)
test reductions_rgba_to_rgb_16              ... bench:   1,911,421 ns/iter (+/- 40,144)
test reductions_rgba_to_rgb_8               ... bench:   1,502,322 ns/iter (+/- 73,882)
```

A practical test of some files at `-o0` showed a speed improvement from 11.5s to 9.8s.

This contributes towards issue #124 - the main thing that might still be done is point 4 on that list. I'm not sure any additional improvements could result in tangible speed gains though, as all the reductions are now a very small fraction of the overall processing time.